requirements.txt
praw
yfinance
schedule
python-dotenv
pandas
requests
